+++ easyfarm/src/main/java/easyfarm/dao/PlanMapper.java
@@ -31,6 +31,44 @@ public interface PlanMapper {
 	public List<TaxPay> getTaxPayPlan();
 	public List<WorkForcePay> getWorkForcePayPlan();
 	
+	//농가별 프로젝트 정보조회
+	public Map<String, Object> getFarmProjectInfo(String projectCode);
+	
+	//프로젝트 계획정보조회
+	public Map<String, Object> getProjectPlanInfo(String projectPlanCode);
+	
+	//프로젝트별 계획차수 리스트 조회
+	public List<Map<String, Object>> getProjectPlanNList(String projectCode);
+	
+	//프로젝트별 max계획차수 조회
+	public Map<String, Object> getMaxProjectPlanNum(String projectCode);
+	
+	//프로젝트 통합계획 등록
+	public int addProjectPlan(Map<String, Object> projectPlanData);
+	
+	//프로젝트별 작업단계
+	public List<Map<String, Object>> getWorkphaseName(Map<String, Object> projectData);
+	
+	//프로젝트별 상세작업항목
+	public List<Map<String, Object>> getWorkphaseCateName(Map<String, Object> projectData);
+	
+	//농가별 거래처
+	public List<Map<String, Object>> getClientName(Map<String, Object> projectData);
+	
+	//농가별 농기계즐겨찾기
+	public List<Map<String, Object>> getFarmBookmarkMachine(Map<String, Object> projectData);
+	
+	//농가별 보유농기계
+	public List<Map<String, Object>> getFarmRetainMachine(Map<String, Object> projectData);
+	
+	//농가별 품목조회
+	public List<Map<String, Object>> getStockItem(Map<String, Object> projectData);
+	
+	//품목정보조회
+	public Map<String, Object> getStockItemInfo(String stockItemCode);
+	
+	//공과금항목조회
+	public List<Map<String, Object>> getTaxPayCateCode();
